api/v1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-arango==7.9.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3 